# docs/api_contract.yaml
openapi: 3.0.3
info:
  title: Resham Sutra WAHA Control Center API
  description: Standalone API documentation for the Resham Sutra WAHA Control Center (WhatsApp Module V1).
  version: 1.0.0
servers:
  - url: http://localhost:3002/api
    description: Local development backend server
paths:
...
  /whatsapp/qr:
    get:
      summary: Get Session QR Code
      description: Requests WAHA session engine to output active QR code link if disconnected.
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Active status QR code.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "SCAN_QR"
                  qrCode:
                    type: string
                    description: Base64 or Text QR output payload
                    example: "2@abc123xyz..."
        '502':
          description: WAHA instance unreachable or connection error.
  /whatsapp/status:
    get:
      summary: Get WhatsApp Session Status
      description: Checks connection status of the WAHA engine and updates MySQL database state.
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Connection status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  session:
                    type: object
                    properties:
                      name:
                        type: string
                        example: default
                      status:
                        type: string
                        example: SCAN_QR
  /whatsapp/disconnect:
    post:
      summary: Disconnect WhatsApp Session
      description: Stops the active WhatsApp session in WAHA and updates MySQL status. Restricted to ADMIN.
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Disconnected successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: WhatsApp session 'default' disconnected.
        '403':
          description: Forbidden. User lacks admin credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: Only administrators can modify WhatsApp connections.
  /whatsapp/restart:
    post:
      summary: Restart WhatsApp Session
      description: Disconnects and restarts the WhatsApp session in WAHA and updates status. Restricted to ADMIN.
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Restarted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: WhatsApp session 'default' restarted.
        '403':
          description: Forbidden. User lacks admin credentials.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: error
                  code:
                    type: string
                    example: FORBIDDEN
                  message:
                    type: string
                    example: Only administrators can modify WhatsApp connections.
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: JWT authentication token set on login.
    WebhookSecretAuth:
      type: http
      scheme: bearer
      description: Secret Bearer token configured for verification.
...